requirements.txt
quart
surrealdb
libsass
pytest
pytest-depends
requests